requirements.txt
PyGithub==1.59.0
pre-commit==3.3.3
